chore: test cleanup
